refactor: state와 action 구분
